Simplify ListFiles action
